requirements.txt
PySide6>=6.6
pyqtgraph>=0.13
numpy>=1.24
pyserial>=3.5
esptool>=4.7
scipy>=1.10
librosa>=0.10
matplotlib>=3.7
scikit-learn>=1.3 